Modified CSS to prevent login button displaying on page load
Unfortunately for whatever reason this isn't working at the moment but, according to my understanding, the login buttons should no longer display until such time as Angular has monkeyed with their CSS display property (it's now set to none by default). I think I'll need to ask for some help to get to the bottom of this on Monday.