Loads friend list
Now loads list of friends and (sort of) displays them in a grid of clickable buttons so you can go to their pages... only there's some bonkersness with bullets and, now that I'm setting the style on the links rather than the list items, the vertical alignment is banjaxxed. It'll do for now!