Added filter box
It's now possible to filter the displayed list of friends by typing in the search box. Unfortunately this shows up one of AngularJS's problems: poor performance when working with larger datasets. It filters on every character and, as you can see if you try it, really grids. I may have to put some delay on the filter later, maybe 500ms or something like that, so it basically only filters once you've stopped typing for a bit.